refactor: Modify Direction
